Get mark as done module working (#2)
* update readme

* run composer install

* fix task seeder, using non existing method

* add system admin to user seeder

* add indicator to navbar item to show if you are on the page

* turn name into span instead of button, and add logout link

* sort by sys admin first then professor

* update wording for start and end labels and update variable naming

* install php-flasher and flash error if kicking your self from a course

* disabled kick button on enrolled member that matches authenticated user id

* add comments in create method of course controller

* fix repository search on local environment

* update readme

* fix: task title not updating

* delete temp files

* mention vue in readme

* fix: color scheme on action bar

* remove dangling file-upload

* fix header for gitlab repo search

* add docs to create group

* fix #126 not showing task names in group view

* fix #117 group tasks redirecting to undefined course id

* update project_id field naming to give more clarity it's related to gitlab and not the model itself

* check if gitlab modules are enabled and only run gitlab code if enabled (fixes not being able to start a text only assignment)

* fix: resetting a project that is not connected to gitlab, still reaches out to gitlab

* fix: undefined variable crash, when having two modules installed, where one is not enabled

* fix headers creation for repo selection

* fix gitlab_project_id not marked as fillable

* remove unnecessary json middleware for repo search

* add toast package to enable giving better user feedback

* fix isOwner check on group information and disallow owner or last person to leave group

* Only allow group owner to delete invites

* set min value for endsAt time, on load and changing startsAt time

* add tooltip to clarify icons in course member management

* add visual feedback when creating/editing a task for when or when not you are allowed to save.

* add database mention in running the application

* run composer update

* run php-cs-fixer

* update readme for running php fixer

* update php version in gha

* setup php for both jobs

* use correct version name

* revert: update project_id field naming to give more clarity it's related to gitlab and not the model itself

* fix tests

* fix some phpstan

* lint file

* establish phpstan baseline

* include phpstan

* Revert "revert: update project_id field naming to give more clarity it's related to gitlab and not the model itself"

This reverts commit dbe30ee022a6c72846788201cf27aeb315f86229.

* fix variable references and naming

* conflict mark as done and link repository

* change method names for task types

* fix mark as done module not passed through to the frontend

* remove mounted method

* run fixer

* add concurrency to github workflows, to reduce usage

* diversify names from same workflow

* add deprecated to correction type enum

* move mark as complete onto project level, and remove grade when resetting project

* unpack variables from functions in blade instead

* make mark as complete test use project id

* add test case when marking a project complete that is not yours

* run php-cs-fixer

* add test case for marking group projects done

* run php-cs-fixer

+++ app/Http/Controllers/ProjectController.php
@@ -34,14 +34,11 @@ use GraphQL\SchemaObject\RepositoryTreeArgumentsObject;
 use GraphQL\SchemaObject\RootProjectsArgumentsObject;
 use GraphQL\SchemaObject\RootQueryObject;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Collection;
 use Illuminate\View\View;
-use Spatie\ShikiPhp\Shiki;
 use Str;
 use Symfony\Component\HttpFoundation\StreamedResponse;
-use function Pest\Laravel\get;
 
 class ProjectController extends Controller
 {
@@ -69,21 +66,51 @@ class ProjectController extends Controller
     {
         abort_unless($project->status == ProjectStatus::Active, 400);
         \DB::transaction(function() use ($gitLabManager, $project) {
-            $found = $project->project_id != null;
+            $found = $project->gitlab_project_id != null;
             try
             {
-                $gitLabManager->projects()->show($project->project_id);
+                // Be aware if passed in value is null, then it will return all projects and therefore not throwing.
+                $gitLabManager->projects()->show($project->gitlab_project_id);
             } catch(RuntimeException $runtimeException)
             {
                 $found = $runtimeException->getCode() != 404;
             }
 
             if($found)
-                $gitLabManager->projects()->remove($project->project_id);
+                $gitLabManager->projects()->remove($project->gitlab_project_id);
 
 
             $project->delete();
+
+            /** @var ?Grade $grade */
+            $grade = Grade::where("task_id", "=", $project->task->id)
+                ->where("source_type", "=", User::class)
+                ->where("source_id", "=", auth()->id())->first();
+            if ($grade)
+            {
+                $grade->delete();
+            }
+
         });
+
+        return "OK";
+    }
+
+    public function markComplete(Course $course, Task $task, Project $project): string|Response
+    {
+        if( ! $task->isTextTask())
+            return response('Bad request', 400);
+        if(Grade::where(['task_id' => $task->id, 'user_id' => auth()->id()])->exists())
+            return response('Bad request', 400);
+
+        $foundProject = $project->owners()->find(auth()->id());
+
+        if ( ! $foundProject)
+        {
+            return response('Forbidden', 403);
+        }
+
+        $project->setProjectStatus(ProjectStatus::Finished);
 
         return "OK";
     }
@@ -119,7 +146,7 @@ class ProjectController extends Controller
         abort_if($sha == null, 404);
 
         return response()->streamDownload(function() use ($sha, $project, $gitLabManager) {
-            echo $gitLabManager->repositories()->archive($project->project_id, [
+            echo $gitLabManager->repositories()->archive($project->gitlab_project_id, [
                 'sha' => $sha,
             ], 'zip');
         }, "$project->repo_name.zip");
@@ -140,7 +167,7 @@ class ProjectController extends Controller
         {
             $rootObject = new RootQueryObject();
             $rootObject->selectProjects((new RootProjectsArgumentsObject())
-                ->setIds(["gid://gitlab/Project/$project->project_id"])
+                ->setIds(["gid://gitlab/Project/$project->gitlab_project_id"])
                 ->setFirst(1))
                 ->selectNodes()
                 ->selectRepository()
